Simple utility to strip IDL docs
The fixed term and margin IDLs are huge because of docs.
We can't upload them locally nor on devnet, and we'll waste SOL on mainnet on all IDLs with docs.

This utility strips out the docs from the JSON files and writes them back.

A correct solution would be to fix this in anchor-lang, however we're now using a released version from crates.io, so it'll take a while for us to pick up this fix.

I'm making this change now, and I'll fix it in anchor when I have some time later.