Fix fixed term valuation
The account summary is incorrect because the fixed term values don't consider USD values.
This would have caused the UI to display incorrect balances for say SOL markets.

The fix is in the data service to return a map of underlying mint and its asset values.
Then the UI converts this to a USD value using latest prices.