tools/alt: add pool and fixed term programs to registry
To maintain some structure, the intention was to also add the fixed term and margin pool positions to their lookup tables so that we can identify them later.

This helps us to create lookup tables that have more structure instead of a random collection of addresses.

We check for the program address when deciding which addresses to append, however as we haven't actually added the addresses, we would have ended up duplicating addresses in the registry.