ui: Include term positions & orders in UI valuation
This replaces the shim that we created as it was meant to be temporary anyway.
I noticed that the shim could get delayed, causing an incorrect balance to be displayed.
It was going to be confusing for users.

In order to move its calculation to the UI, we:
* Started consuming open orders and positions that were already being emitted by ws
* Added a margin account's orders and positions in its constructor and valuation methods
* Made a small change to expose the underlying mint of positions, to avoid looking them up on the market each time

This doesn't yet replace the REST calls that we use to get the same information for the tables in the fixed term pages.
We'll follow up with that change.