ui: attempt 2 at reconnects
The problem with the initial implementation of reconnects was that even when we close the connection, we'd restart the same connection, causing memory leaks.

After reading the spec, I found that a saner solution is to signal a code when disconnecting.
We then check for this code on `ws.onclose` and don't reconnect if we closed by that code.
This will allow us to recover true connection failures.