ui: don't try send single versioned tx
While we don't yet have margin account lookup tables, users with a lot of position will still reach the tx size limit.

This change splits the tx into 2, the first to refresh positions, then the next to perform an action.

This is only on fixed-term, we hadn't started using versioned tx on other interactions.