ui: upgrade margin account by adding lookup tables
When a user selects a margin account, we show them a modal to upgrade
the account if applicable.
We use a threshold of 10 accounts, where if they need to add 10 or more
accounts, they have to upgrade.
The user can dismiss the upgrade, however they'll see the popup next time
they refresh the page or select the margin account.

This is probably fine because we do want users to upgrade their accounts as
otherwise their experience can be adversely affected.

We show the user what actions are being performed, so we can use this interface
in future for cleaning up or some other action that has to be taken by the margin account.