ui: fix incorrect pool repayment constraint
When a user is repaying a pool loan, the maximum they can repay should be the lower of the loan and their deposit.

The maximum was currently set to their wallet balance, which was incorrectly preventing users from repaying their debts on the UI.

The risk forecast is still incorrect, it will be addressed separately.